+++ tests/__init__.py
@@ -1,2 +1,2 @@
-async def noop() -> None:
+async def do_absolutely_nothing() -> None:
     """Do nothing."""